requirements.txt
paho-mqtt
flask
shapely
